requirements.txt
numpy==1.24.3
torch
openai-whisper
pydub
ffmpeg-python
certifi